requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
build==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
dateparser==1.2.2
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
multidict==6.6.4
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pluggy==1.6.0
propcache==0.3.2
pycryptodome==3.23.0
Pygments==2.19.2
pyproject_hooks==1.2.0
pytest==8.4.1
python-binance==1.0.29
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.7.34
requests==2.32.5
retrying==1.4.2
six==1.17.0
TA-Lib==0.6.5
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
